fix 271

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/util/bloom/BloomFilter.java
@@ -234,6 +234,6 @@ public class BloomFilter extends Filter {
   
   /* @return number of bytes needed to hold bit vector */
   private int getNBytes() {
-    return (int)(((long)vectorSize + 7) / 8);
+    return (vectorSize + 7) / 8;
   }
 }//end class
